requirements.txt
requests
python-dotenv
pandas
rich
matplotlib
seaborn
scikit-learn